+++ packages/widgetbook/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - **FEAT**: Add [Semantics Addon](https://docs.widgetbook.io/addons/semantics-addon). ([#1402](https://github.com/widgetbook/widgetbook/pull/1402))
+- **FEAT**: Add a banner to the bottom of the navigation panel that shows the total number of components and use-cases. ([#1405](https://github.com/widgetbook/widgetbook/pull/1405))
 
 ## 3.12.0